클래스 책임 분리
항상 한 클래스가 하나의 책임을 맡도록 하는 것이 바람직하다.
- 권한부여
- 사용자 관리 
로 구성을 분리했다.
그리고 그에 맞게 테스트 코드 @Improt를 수정했다.
