+++ src/main/java/com/kurt/practice/springbootsecuritypractice/config/UserManagementConfig.java
@@ -1,29 +1,17 @@
 package com.kurt.practice.springbootsecuritypractice.config;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.NoOpPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
 
-public class SecurityConfig {
-
-    @Bean
-    protected SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .authorizeHttpRequests(authz ->
-                        authz.anyRequest().authenticated())
-                .httpBasic();
-
-        return http.build();
-    }
+public class UserManagementConfig {
 
     @Bean
     public UserDetailsService userDetailsService() {
-         UserDetailsService userDetailsService = new InMemoryUserDetailsManager();
+        UserDetailsService userDetailsService = new InMemoryUserDetailsManager();
 
         User user = (User) User.withUsername("kurt")
                 .password("12345")
@@ -39,6 +27,5 @@ public class SecurityConfig {
     public PasswordEncoder passwordEncoder() {
         return NoOpPasswordEncoder.getInstance();
     }
-
 
 }
